Use the framebuffer resolution in safe|default if it is "good"
The definition of "good" is somewhat subjective and this may be
a little dangerous (especially for "safe" mode) but it means we
will try to make available the same resolution they have selected
for the frame buffer if it is not 1280x1024, 1024x768, or
800x600 and if the width is greater than 1024.

The framebuffer resolution was selected by the user and by reading
it from /sys/.../virtual_size we know the hardware thinks it is
a viable resolution.  In addition we still provide the fallbacks
of 1280x1024, etc.

Maybe this will be a reasonable compromise.  It would also free
us up to make the default default size less than 1280x1024 (but
I would have to tune good_fb_res() slightly so 1280x1024 is
selected if it is used in the fb.